requirements.txt
attrs==17.4.0
Babel==2.5.1
certifi==2018.1.18
chardet==3.0.4
coverage==4.5
decorator==4.2.1
Django==1.11.9
django-countries==5.0
django-extra-views==0.6.4
django-haystack==2.6.1
django-localflavor==2.0
django-mailgun==0.9.1
django-oscar==1.5.1
django-phonenumber-field==1.3.0
django-tables2==1.17.1
django-treebeard==4.2.0
django-widget-tweaks==1.4.1
factory-boy==2.9.2
Faker==0.8.8
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
mixer==6.0.1
mock==2.0.0
mysqlclient==1.3.12
parso==0.1.1
pbr==3.1.1
pexpect==4.3.1
phonenumberslite==8.8.9
pickleshare==0.7.4
Pillow==5.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
purl==1.3.1
py==1.5.2
pycountry==17.9.23
Pygments==2.2.0
pytest==3.4.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-ipdb==0.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
sorl-thumbnail==12.4.1
text-unidecode==1.1
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
WebOb==1.7.4
